Member Force Refactor
In some cases the .gto file can contain legitimate warnings. These warnings alter the spacing and formatting of the member results sets when they occur. To resolve this the way the member forces were parsed has been overhauled. The member results set is no longer read line by line. Instead it is read once to split the results into 'blocks' of members. Then each block is checked for warnings and number of load combinations and joints.