Overhaul of user data selection storage and interface generation
    ** currently does not generate results but the interface development has been streamlined ** - There is light at the end of the tunnel for this change, a very dim light.

Added a results parameters class which contains a dictionary of all selected user results which can then be  accessed at any time

Developed a configuration file to store internal data which is used multiple times such as headings and file types such that any updates can occur in a single location and will propagate throughout the files.

Unified the generation of the widgets on the tabs - consolidated to single file instead of copied 3 times.