bug fixes for user input list items
Fixed errors associated with a user selecting to input a list of requested values. List parsing was not producing the correct results when both the name and load spec were lists.

Similarly, the error notification would not catch multiple errors in the list, it would only catch the last error.

Both bugs have been fixed.