bug fixes for user input list items (#20)
Fixed errors associated with a user selecting to input a list of requested values. List parsing was not producing the correct results when both the name and load spec were lists.

Similarly, the error notification would not catch multiple errors in the list, it would only catch the last error.

Finally, if the user deleted all of the existing results sets the results.selection dictionary would be empty, meaning any new values would not be 'replaced'. Changed default dictionary to empty and changed methodology to use append if the length of the results sets was zero.

All three bugs have been fixed.