bug fixes and interface improvement (#21)
Fixed that if an interaction ratio range was entered as text (i.e. string) an error would not be returned. Now, a valid input is required or the user is notified of an error.

Fixed scroll bar in available_results_tree. The scroll bar was defualt size, and while it would scroll it would not change size to match the number of items in the available_results_tree.

Changed font of available results and selected results to stand out.

Greyed out IR input boxes when the text regions were disabled. Will likely extend to other text boxes in future.