sorting and data storage bug fixes
Fixed sorting, was using the wrong list when doing multisort. Basically just sorting the initial list for criteria A, and then resorting per criteria B instead of sorting the new list created by criteria A.

Fixed 'none' type object for input file path due to reinitializing the ResultsParameters class. Built a proper reset function.

Fixed that the sort order (asc. or dec.) was not stored when generating a properties file.

Fixed that the sort window would not open when modifying a previously generated result which had sorting as 'True' check box was checked correctly, however the sort window wasn't called.