Minor improvement, code cleanup, bug fixes.
Fixed that checking for IR errors returned invalid IR error incorrectly. This occurred when either greater than or less than was chosen because the tuple contained a blank ' ', which always failed the try / except block. Updated code to ignore ' ' since it is a valid selection.

Fixed error where if the user had an output file with a commented out 'LIST FORCE' it would appear in the available results window even though it is not valid due to being commented out. Modified program to ignore LIST FORCE commands if there is a '$' in the line.

Added '---' to the error handling window for results parameters that are valid when an error is found in the set.

Updated paths to user guide, and any images/icons to be relative to the installed location

added images/icons/user guide to repository