Fixed minor bugs based on preliminary tester feedback
Fixed that the available results column headers would not resize on double click

Fixed that when pressing the back button, the continue button was active but would not actual do anything. This is resolved by clearing out the file and directory text boxes when the back button is pressed.

fixed that the user input was case sensitive. All user inputs are now converted to upper case.

added one more testing case.