Bug fixes based on testing new .gto file (#33)
Fixed bug where if 'output by member' was the very first line of the file an error would be returned because the line number was 0 (or 'none' or 'false') thus showing up as not existing. Changed code to validate that the string_position is not None

added splash screen to executable with text for loading and 'complete' based on a sleep time.

updated user guide to latest version

added splash image file

fixed error where if there was a warning message at the top of the code check the warning message would be displayed instead of the input line number. There were no issues with the actual operation of the program, just the incorrect input file line was displayed in the available results window. This was solved by finding the 'trigger_string' and then looping backwards through the input file from that point until the program found '{' indicating the correct input line number.