tested additional output files and found some issues to fix
fixes that the amount of lines between the start of the code check and the user input line can vary. Program finds the important line then iterates through the elements until it fines the start of the output file

Fixed that checking code check column 1 for specific code names doesn't do anything

standardized how the headers are extracted from member forces and joint reactions wit re.split

updated the parse file for data function to capture cases where the user command is 'LIST FORCES REACTIONS' instead of separate commands for forces and reactions. Previously the code was unable to execute if the command was on the same input line.

removed the need for a generic 'skip lines' value telling the program how many lines to skip. One magic number gone!

Simplified the utilities function for displaying the available results values.