Update check_input_errors.py
Created standalone methods for sections of code that were repeated multiple times.

Added method for checking if a user criteria is valid, returns true if it's valid and false if it is not.

Added method for returning a list of invalid user criteria

Replaced the duplicated code with the two previously described methods.