Reformatted member forces to use block objects (#39)
Started the process of reformatting the code to use Classes for each result block. This commit adds the base class and the member force sub class.

Additionally, the extract_member_forces file was restructured to use the sub class and several functions were split out into a new 'avialable_results_tools' module. This module will be reused for both the member forces and code check refactor in the future.