updated code check module to use common methods
Revised the code check model to use common functions like valid_names for returning a list of valid beam names and profile names in the code check

streamlined the IR parsing function

created separate method for grabbing the contents of a column when represented as a nested list.

streamline the sorting and output building methods to use fewer variables since all the information was already contained in existing variables.